requirement fulfilled

+++ backend/src/database/default/repository/shiftRepository.ts
@@ -4,6 +4,9 @@ import {
   FindOneOptions,
   FindConditions,
   DeleteResult,
+  Between,
+  MoreThanOrEqual,
+  LessThanOrEqual,
 } from "typeorm";
 import moduleLogger from "../../../shared/functions/logger";
 import Shift from "../entity/shift";
@@ -14,7 +17,15 @@ const logger = moduleLogger("shiftRepository");
 export const find = async (opts?: FindManyOptions<Shift>): Promise<Shift[]> => {
   logger.info("Find");
   const repository = getRepository(Shift);
-  const data = await repository.find(opts);
+  const { where, order } = opts;
+  const startDate = where[0];
+  const endDate = where[1];
+  const data = await repository.find({
+    where: {
+      date: Between(startDate, endDate),
+    },
+    order,
+  });
   return data;
 };
 
@@ -38,9 +49,45 @@ export const findOne = async (
   return data;
 };
 
-export const create = async (payload: Shift): Promise<Shift> => {
+export const create = async (payload: any): Promise<Shift> => {
   logger.info("Create");
   const repository = getRepository(Shift);
+  let { date, startTime, endTime } = payload;
+  const first = date.getDate() - date.getDay() + 1;
+  const last = first + 6;
+  const monday = new Date(date.setDate(first));
+  const sunday = new Date(date.setDate(last));
+  const publishChecker = await repository.find({
+    where: {
+      date: Between(monday, sunday),
+    },
+  });
+  if (publishChecker.length) {
+    if (publishChecker[0].isPublished) {
+      return;
+    }
+  }
+  const data = await repository.find({
+    where: [
+      {
+        startTime: Between(startTime, endTime),
+        endTime: Between(startTime, endTime),
+        date,
+      },
+      {
+        endTime: Between(startTime, endTime),
+        date,
+      },
+      {
+        startTime: LessThanOrEqual(startTime),
+        endTime: MoreThanOrEqual(endTime),
+        date,
+      },
+    ],
+  });
+  if (data.length) {
+    return;
+  }
   const newdata = await repository.save(payload);
   return newdata;
 };
@@ -51,14 +98,66 @@ export const updateById = async (
 ): Promise<Shift> => {
   logger.info("Update by id");
   const repository = getRepository(Shift);
+  const { date, startTime, endTime } = payload;
+  const data = await repository.find({
+    where: [
+      {
+        startTime: Between(startTime, endTime),
+        endTime: Between(startTime, endTime),
+        date,
+      },
+      {
+        endTime: Between(startTime, endTime),
+        date,
+      },
+      {
+        startTime: LessThanOrEqual(startTime),
+        endTime: MoreThanOrEqual(endTime),
+        date,
+      },
+    ],
+  });
+  if (data.length) {
+    return;
+  }
+  const shift = await repository.findOne({
+    id,
+  });
+  if (shift.isPublished) {
+    return;
+  }
   await repository.update(id, payload);
   return findById(id);
 };
 
-export const deleteById = async (
-  id: string | string[]
-): Promise<DeleteResult> => {
+export const deleteById = async (id: any): Promise<DeleteResult> => {
   logger.info("Delete by id");
   const repository = getRepository(Shift);
+  const shift = await repository.findOne({
+    id,
+  });
+  if (shift.isPublished) {
+    return;
+  }
   return await repository.delete(id);
+};
+
+export const publishShift = async (
+  opts?: FindManyOptions<Shift>
+): Promise<Shift[]> => {
+  logger.info("Publish");
+  const repository = getRepository(Shift);
+  const { where } = opts;
+  const startDate = where[0];
+  const endDate = where[1];
+  await repository
+    .createQueryBuilder()
+    .update({
+      isPublished: true,
+    })
+    .where({
+      date: Between(startDate, endDate),
+    })
+    .execute();
+  return find(opts);
 };
